aiit

+++ delete.php
@@ -1,4 +1,6 @@
 <?php
+session_start();
+$user_id=$_SESSION['id'];
 /**
  * Created by PhpStorm.
  * User: Kollybright
@@ -12,7 +14,7 @@ include("conn.php");
 $id=$_POST['the'];
 if ($id){
 
-    $sql= "DELETE FROM  todo WHERE event_id=$id";
+    $sql= "DELETE FROM  todo WHERE event_id='".$id."'  AND user_id='".$user_id."'" ;
      $query = mysqli_query($link,$sql);
      if ($query){
          echo "reload";
